refactor: remove deprecated security configuration classes, update references to `camunda-security-library` equivalents
