+++ qa/acceptance-tests/src/test/java/io/camunda/it/csrf/CsrfTokenIT.java
@@ -16,7 +16,7 @@ import io.camunda.qa.util.cluster.TestCamundaApplication;
 import io.camunda.qa.util.cluster.TestWebappClient.TestLoggedInWebappClient;
 import io.camunda.qa.util.multidb.MultiDbTest;
 import io.camunda.qa.util.multidb.MultiDbTestApplication;
-import io.camunda.security.configuration.CsrfConfiguration;
+import io.camunda.security.api.model.config.CsrfConfiguration;
 import java.io.IOException;
 import java.net.URISyntaxException;
 import java.net.http.HttpClient;
