+++ security/security-core/src/main/java/io/camunda/security/configuration/SecurityConfiguration.java
@@ -8,6 +8,10 @@
 package io.camunda.security.configuration;
 
 import io.camunda.security.api.model.config.AuthenticationConfiguration;
+import io.camunda.security.api.model.config.AuthorizationsConfiguration;
+import io.camunda.security.api.model.config.CsrfConfiguration;
+import io.camunda.security.api.model.config.MultiTenancyConfiguration;
+import io.camunda.security.api.model.config.SaasConfiguration;
 import io.camunda.security.api.model.config.headers.HeaderConfiguration;
 import java.util.regex.Pattern;
