+++ zeebe/gateway-rest/src/main/java/io/camunda/zeebe/gateway/rest/controller/ConditionalController.java
@@ -11,7 +11,7 @@ import io.camunda.gateway.mapping.http.RequestMapper;
 import io.camunda.gateway.mapping.http.ResponseMapper;
 import io.camunda.gateway.protocol.model.ConditionalEvaluationInstruction;
 import io.camunda.security.api.context.CamundaAuthenticationProvider;
-import io.camunda.security.configuration.MultiTenancyConfiguration;
+import io.camunda.security.api.model.config.MultiTenancyConfiguration;
 import io.camunda.service.ConditionalServices;
 import io.camunda.service.ConditionalServices.EvaluateConditionalRequest;
 import io.camunda.zeebe.gateway.rest.annotation.CamundaPostMapping;
