+++ zeebe/gateway-rest/src/main/java/io/camunda/zeebe/gateway/rest/controller/ResourceController.java
@@ -16,7 +16,7 @@ import io.camunda.gateway.protocol.model.ResourceSearchQuery;
 import io.camunda.gateway.protocol.model.ResourceSearchQueryResult;
 import io.camunda.search.query.DeployedResourceQuery;
 import io.camunda.security.api.context.CamundaAuthenticationProvider;
-import io.camunda.security.configuration.MultiTenancyConfiguration;
+import io.camunda.security.api.model.config.MultiTenancyConfiguration;
 import io.camunda.service.ResourceServices;
 import io.camunda.service.ResourceServices.DeployResourcesRequest;
 import io.camunda.service.ResourceServices.ResourceDeletionRequest;
