+++ zeebe/gateway-rest/src/main/java/io/camunda/zeebe/gateway/rest/controller/SignalController.java
@@ -12,7 +12,7 @@ import io.camunda.gateway.mapping.http.RequestMapper.BroadcastSignalRequest;
 import io.camunda.gateway.mapping.http.ResponseMapper;
 import io.camunda.gateway.protocol.model.SignalBroadcastRequest;
 import io.camunda.security.api.context.CamundaAuthenticationProvider;
-import io.camunda.security.configuration.MultiTenancyConfiguration;
+import io.camunda.security.api.model.config.MultiTenancyConfiguration;
 import io.camunda.service.SignalServices;
 import io.camunda.zeebe.gateway.rest.annotation.CamundaPostMapping;
 import io.camunda.zeebe.gateway.rest.mapper.RequestExecutor;
