+++ zeebe/gateway-rest/src/test/java/io/camunda/zeebe/gateway/rest/controller/FaultyControllerTest.java
@@ -10,7 +10,7 @@ package io.camunda.zeebe.gateway.rest.controller;
 import static org.mockito.Mockito.when;
 
 import io.camunda.security.api.context.CamundaAuthenticationProvider;
-import io.camunda.security.configuration.MultiTenancyConfiguration;
+import io.camunda.security.api.model.config.MultiTenancyConfiguration;
 import io.camunda.service.ProcessInstanceServices;
 import io.camunda.zeebe.gateway.rest.RestControllerTest;
 import org.junit.jupiter.api.BeforeEach;
